DSND-541:Address review comments
