+++ src/main/java/uk/gov/companieshouse/charges/data/converter/ReadConverter.java
@@ -1,24 +1,24 @@
 package uk.gov.companieshouse.charges.data.converter;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
-import com.mongodb.BasicDBObject;
+import org.bson.Document;
 import org.springframework.core.convert.converter.Converter;
-import org.springframework.data.convert.WritingConverter;
+import org.springframework.data.convert.ReadingConverter;
 import uk.gov.companieshouse.api.charges.ChargeApi;
 
-@WritingConverter
-public class CompanyChargesWriterConverter implements Converter<ChargeApi, BasicDBObject> {
+@ReadingConverter
+public class ReadConverter implements Converter<Document, ChargeApi> {
 
     private final ObjectMapper objectMapper;
 
-    public CompanyChargesWriterConverter(ObjectMapper objectMapper) {
+    public ReadConverter(ObjectMapper objectMapper) {
         this.objectMapper = objectMapper;
     }
 
     @Override
-    public BasicDBObject convert(ChargeApi source) {
+    public ChargeApi convert(Document source) {
         try {
-            return BasicDBObject.parse(objectMapper.writeValueAsString(source));
+            return objectMapper.readValue(source.toJson(), ChargeApi.class);
         } catch (Exception ex) {
             throw new RuntimeException(ex);
         }
